refactre nonui
